Setup JavaScript -- very simplified, no REPL.
This will do the basics, but I don't know how (yet) to configure
nodejs-repl to use the correct version of node via nvm. Also, I'm not
using tern because it's not currently maintained and doesn't offer a
ton of useful functionality. It would be good for completions, but
it's okay for now. I may try Indium.